appdescription.md
Final version of the app. Keeps tracks of the statistics of a football game such as goals scored , fouls commited and etc.. Available in 2 languages : Portuguese and English (GB). Includes a reset button that resets the scores. App is only available in portrait mode. Also has customized buttons.Made on a 768x1280 xhdpi Nexus 4 resolution with Android Studio.